don't remove the _bookdown_files folder by default, as it stores files used for caching

+++ scripts_and_filters/knit-functions.R
@@ -1,4 +1,4 @@
-knit_thesis <- function(input, output_format = "pdf", ...){
+knit_thesis <- function(input, output_format = "pdf", allow_cache = TRUE, ...){
   if ("pdf" %in% output_format){
     bookdown::render_book(input, output_format = "bookdown::pdf_book")
     
@@ -23,6 +23,8 @@ knit_thesis <- function(input, output_format = "pdf", ...){
     bookdown::render_book(input, output_format = "bookdown::word_document2")
   }
   
-  # remove the _bookdown_files folder after the build
-  unlink("_bookdown_files", recursive = TRUE)
+  if (!allow_cache){
+    # remove the _bookdown_files folder after the build
+    unlink("_bookdown_files", recursive = TRUE)  
+  }
 }
